+++ app/_layout.tsx
@@ -5,6 +5,14 @@ export default function RootLayout() {
         <Stack screenOptions={{ headerShown: false }}>
             <Stack.Screen name="login" />
             <Stack.Screen name="(tabs)" />
+            <Stack.Screen
+                name="details"
+                options={{
+                    headerShown: true,
+                    title: "Détails",
+                    headerBackTitle: "Retour",
+                }}
+            />
         </Stack>
     );
 }