+++ server/src/server.ts
@@ -1,0 +1,17 @@
+import dotenv from "dotenv";
+import express from "express";
+import path from "path";
+import { fileURLToPath } from "url";
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+dotenv.config({ path: path.resolve(__dirname, "../../.env") });
+
+const PORT_BACKEND = process.env.PORT_BACKEND;
+const ORIGIN_URL = process.env.ORIGIN_URL;
+const SERVER_URL = process.env.SERVER_URL;
+
+const app = express();
+
+app.listen(PORT_BACKEND, () => {
+    console.log(`Server is running on ${SERVER_URL}`);
+});